Bugfix/visible form adding player in team (#518) (#523)
* fix: Correct misprints in template.
fix: Fix visible form for adding teams.
fix: Save added or updated player's teams.

* fix: Change PlayerForm for visible team form.

* fix: Move mixin in mixins.py and ajax view in ajax.py.
